refactor omise error
