+++ OmiseSDK/Sources/Models/OmiseErrors/OmiseError+LocalizedDescription.swift
@@ -425,4 +425,3 @@ extension OmiseError {
         }
     }
 }
-// swiftlint:enable line_length
