+++ OmiseSDK/Sources/Views/Screens/Choose Payment Methods/Select Payment From Lists/SelectPaymentController.swift
@@ -29,7 +29,7 @@ class SelectPaymentController: UITableViewController {
         }
     }
 
-    @objc private func didTapClose() {
+    @objc func didTapClose() {
         viewModel.viewDidTapClose()
     }
